delay_api/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.8.1
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
Django==5.0.6
django-redis==5.4.0
djangorestframework==3.15.1
fastjsonschema==2.19.1
frozenlist==1.4.1
hiredis==2.3.2
idna==3.7
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_core==5.7.2
multidict==6.0.5
nbformat==5.10.4
pigar==2.1.4
platformdirs==4.2.2
psycopg==3.1.19
pymemcache==4.0.0
redis==5.0.6
referencing==0.35.1
requests==2.32.3
rpds-py==0.18.1
sqlparse==0.5.0
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.1
yarl==1.9.4
